requirements.txt
opencv-python
numpy
matplotlib
scikit-learn
scikit-image
joblib
pillow
